requirements.txt
pybit==5.6.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.14.0
keras==2.14.0
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.4
ta==0.10.2
talib-binary==0.4.28
pyyaml==6.0.1
aiohttp==3.9.1
asyncio==3.4.3
psutil==5.9.6
Flask==3.0.0
Flask-CORS==4.0.0
Flask-HTTPAuth==4.8.0
Werkzeug==3.0.1
